backend/requirements.txt
fastapi==0.112.0
PyJWT==2.9.0
pymongo==4.6.3
uvicorn==0.30.6
